Make Systemnote's of the Logs_IP_Actions User-friendly
Requesting to make the following changes to the system notes in the logs_ip_actions table to make more sense (especially to special people). You may be wondering why I chose "Logged into WoW", this is because if you have a website for your server that also makes use of this table it will avoid confusion as to whether it was the account of the forum user or his server account. I can change it to "Logged into Account" if more preferred.